backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
sympy>=1.12.0
pydantic>=2.4.0
